Fix/small issues (#364)
* fixed small text

close #363

* moved logout to bottom

* avoid restarting app when logging out if no credentials are found

close #360

* updated libraries

* bumped version for release

* updated workflow

+++ app/app/src/main/java/com/github/livingwithhippos/unchained/settings/viewmodel/SettingsViewModel.kt
@@ -58,12 +58,19 @@ constructor(
 
     fun userLogout() {
         viewModelScope.launch {
-            protoStore.deleteCredentials()
-            eventLiveData.postEvent(SettingEvent.Logout)
+            val credentials = protoStore.getCredentials()
+            if (credentials.accessToken.isBlank() && credentials.clientId.isBlank()) {
+                eventLiveData.postEvent(SettingEvent.LogoutNoCredentials)
+            } else {
+                protoStore.deleteCredentials()
+                eventLiveData.postEvent(SettingEvent.Logout)
+            }
         }
     }
 }
 
 sealed class SettingEvent {
-    object Logout : SettingEvent()
+    data object Logout : SettingEvent()
+
+    data object LogoutNoCredentials : SettingEvent()
 }
